optimized with integer fixed point maths
-int32_t's in move_tool_abs now uses integer math to avoid long floating
point calculations and to increase acurracy
-step_delay doesnt uses acceleration unless more than 10 steps, other
wise rounding errors cause no delay and the motor skips steps
-host software requires modifications to fix com bugs
-checks thermistor temperatures less frequently

+++ replicator_uno/step_delay_micro.ino
@@ -2,7 +2,7 @@
 properly makes a delay from a uint32 amount of microseconds
 ------------------*/
 
-void step_delay_micro(uint32_t _del){
+void step_delay_micro(unsigned long _del){
   delay((_del - (_del % 1000))/1000);
   if ((_del % 1000) > 0){
     delayMicroseconds(_del % 1000);
